pyproject.toml
[project]
name = "img2midi"
version = "0.1.0"
description = "Convert images to MIDI files"
requires-python = ">=3.9"
dependencies = ["pillow>=10.0", "mido>=1.3", "numpy>=1.24"]

[project.scripts]
img2midi = "img2midi:main"

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["img2midi"]
